refactor: 폴더구조 단순화. FSD 제거
