--- src/refactoring/entities/cart/index.ts
@@ -1,3 +1,0 @@
-export * from "./lib";
-export * from "./types";
-export * from "./model";
