--- src/refactoring/entities/cart/types/index.ts
@@ -1,6 +1,0 @@
-import { Product } from "@r/entities/product";
-
-export interface CartItem {
-  product: Product;
-  quantity: number;
-}
