--- src/refactoring/entities/coupon/index.ts
@@ -1,3 +1,0 @@
-export * from "./model";
-export * from "./types";
-export * from "./lib";
