--- src/refactoring/entities/coupon/lib/index.ts
@@ -1,10 +1,0 @@
-import { Coupon } from "@r/entities/coupon";
-
-export const getCouponAppliedTotal = (amount: number, coupon: Coupon) => {
-  switch (coupon.discountType) {
-    case "amount":
-      return Math.max(0, amount - coupon.discountValue);
-    case "percentage":
-      return amount * (1 - coupon.discountValue / 100);
-  }
-};
