--- src/refactoring/entities/product/lib/index.ts
@@ -1,5 +1,0 @@
-import { Discount } from "../types";
-
-export const getMaxDiscount = (discounts: Discount[]) => {
-  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-};
