--- src/refactoring/features/product-list/index.ts
@@ -1,2 +1,0 @@
-export { ProductList } from "./ui";
-export { getRemainingStock } from "./lib";
