--- src/refactoring/hooks/use-cart.ts
@@ -1,45 +1,0 @@
-// useCart.ts
-import { useState } from "react";
-import { CartItem, Coupon, Product } from "@/types";
-import {
-  calculateCartTotal,
-  updateCartItemQuantity,
-} from "@r/entities/cart/lib";
-
-export const useCart = () => {
-  const [cart, setCart] = useState<CartItem[]>([]);
-  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
-
-  const addToCart = (product: Product) => {
-    setCart((prev) => {
-      const existingItem = prev.find((p) => p.product.id === product.id);
-      return existingItem
-        ? updateCartItemQuantity(prev, product.id, existingItem.quantity + 1)
-        : [...prev, { product, quantity: 1 }];
-    });
-  };
-
-  const removeFromCart = (productId: string) => {
-    setCart((prev) => prev.filter((p) => p.product.id !== productId));
-  };
-
-  const updateQuantity = (productId: string, newQuantity: number) => {
-    setCart((prev) => updateCartItemQuantity(prev, productId, newQuantity));
-  };
-
-  const applyCoupon = (coupon: Coupon) => {
-    setSelectedCoupon(coupon);
-  };
-
-  const calculateTotal = () => calculateCartTotal(cart, selectedCoupon);
-
-  return {
-    cart,
-    addToCart,
-    removeFromCart,
-    updateQuantity,
-    applyCoupon,
-    calculateTotal,
-    selectedCoupon,
-  };
-};
