+++ src/refactoring/model/cart/types.ts
@@ -1,0 +1,6 @@
+import { Product } from "../product/types";
+
+export interface CartItem {
+  product: Product;
+  quantity: number;
+}
