+++ src/refactoring/model/discount/lib.ts
@@ -1,0 +1,5 @@
+import { Discount } from "./types";
+
+export const getMaxDiscount = (discounts: Discount[]) => {
+  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+};
