+++ src/refactoring/model/discount/types.ts
@@ -1,0 +1,4 @@
+export interface Discount {
+  quantity: number;
+  rate: number;
+}
