+++ src/refactoring/model/product/lib.ts
@@ -1,0 +1,7 @@
+import { CartItem } from "../cart/types";
+import { Product } from "./types";
+
+export const getRemainingStock = (product: Product, cart: CartItem[]) => {
+  const cartItem = cart.find((item) => item.product.id === product.id);
+  return product.stock - (cartItem?.quantity || 0);
+};
