+++ src/refactoring/model/product/types.ts
@@ -1,0 +1,9 @@
+import { Discount } from "../discount/types";
+
+export interface Product {
+  id: string;
+  name: string;
+  price: number;
+  stock: number;
+  discounts: Discount[];
+}
