+++ src/refactoring/model/product/use-product.ts
