--- src/refactoring/widget/cart-list/index.ts
@@ -1,1 +1,0 @@
-export { CartList } from "./ui";
