--- src/refactoring/widget/coupon-applier/index.ts
@@ -1,1 +1,0 @@
-export { CouponApplier } from "./ui";
