--- src/refactoring/widget/coupon-management/index.ts
@@ -1,1 +1,0 @@
-export { CouponManagement } from "./ui";
