--- src/refactoring/widget/index.ts
@@ -1,5 +1,0 @@
-export * from "./cart-list";
-export * from "./product-management";
-export * from "./coupon-applier";
-export * from "./order-summary";
-export * from "./coupon-management";
