--- src/refactoring/widget/order-summary/index.ts
@@ -1,1 +1,0 @@
-export { OrderSummary } from "./ui";
