--- src/refactoring/widget/product-management/index.ts
@@ -1,1 +1,0 @@
-export { ProductManagement } from "./ui";
